backend/requirements.txt
# FastAPI Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.2.5

# LLM & AI
langchain==0.1.6
langchain-anthropic==0.1.1
langchain-community==0.0.19
anthropic==0.18.1
tiktoken==0.6.0

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.4
unstructured==0.12.4

# Vector Embeddings
sentence-transformers==2.3.1
numpy==1.26.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
